# docs/swagger.yaml
definitions:
...
  models.CustomsDocument:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      document_number:
        type: string
      document_type:
        description: COMMERCIAL_INVOICE, PACKING_LIST, BOL, CUSTOMS_DECLARATION, CERTIFICATE_OF_ORIGIN
        type: string
      document_url:
        type: string
      expiry_date:
        type: string
      id:
        type: integer
      issued_date:
        type: string
      issuing_authority:
        type: string
      load_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Manifest:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      destination_country:
        type: string
      document_url:
        type: string
      generated_at:
        type: string
      id:
        type: integer
      load_count:
        type: integer
      manifest_number:
        type: string
      origin_country:
        type: string
      total_value:
        type: number
      total_volume:
        type: number
      total_weight:
        type: number
      trip_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  title: Triplink API
  version: "1.0"
paths:
...
  /customs-documents:
    post:
      consumes:
      - application/json
      description: Create a new customs document for a load
      parameters:
      - description: Customs document data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/models.CustomsDocument'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Create customs document
      tags:
      - customs
  /customs-documents/{id}:
    delete:
      description: Delete a customs document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete customs document
      tags:
      - customs
    get:
      description: Get a specific customs document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Get customs document by ID
      tags:
      - customs
    put:
      consumes:
      - application/json
      description: Update an existing customs document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated document data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/models.CustomsDocument'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Update customs document
      tags:
      - customs
  /loads/{load_id}/bill-of-lading:
    post:
      description: Generate a bill of lading for a load
      parameters:
      - description: Load ID
        in: path
        name: load_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Generate bill of lading
      tags:
      - customs
  /loads/{load_id}/commercial-invoice:
    post:
      description: Generate a commercial invoice for a load
      parameters:
      - description: Load ID
        in: path
        name: load_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Generate commercial invoice
      tags:
      - customs
  /loads/{load_id}/customs-documents:
    get:
      description: Get all customs documents for a specific load
      parameters:
      - description: Load ID
        in: path
        name: load_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CustomsDocument'
            type: array
      summary: Get customs documents for a load
      tags:
      - customs
  /loads/{load_id}/packing-list:
    post:
      description: Generate a packing list for a load
      parameters:
      - description: Load ID
        in: path
        name: load_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomsDocument'
      summary: Generate packing list
      tags:
      - customs
...
  /manifests/{id}:
    get:
      description: Get a specific manifest by its ID
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manifest'
      summary: Get manifest by ID
      tags:
      - manifests
  /manifests/{id}/detailed:
    get:
      description: Get manifest with detailed load information
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get detailed manifest data
      tags:
      - manifests
  /manifests/{id}/document:
    put:
      consumes:
      - application/json
      description: Update the document URL for a manifest (after PDF generation)
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document URL
        in: body
        name: data
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manifest'
      summary: Update manifest document URL
      tags:
      - manifests
...
  /trips/{trip_id}/customs-summary:
    get:
      description: Get a summary of all customs documents for loads in a trip
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get customs documents summary for trip
      tags:
      - customs
  /trips/{trip_id}/manifest:
    get:
      description: Get the manifest for a specific trip
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manifest'
      summary: Get manifest by trip ID
      tags:
      - manifests
    post:
      description: Generate a consolidated manifest for all loads in a trip
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Manifest'
      summary: Generate manifest for a trip
      tags:
      - manifests
...
swagger: "2.0"